+++ climada_petals/entity/exposures/spam_agrar.py
@@ -24,6 +24,7 @@ import zipfile
 from pathlib import Path
 import pandas as pd
 import numpy as np
+from shapely.geometry import Point
 
 from climada import CONFIG
 from climada.entity.exposures.base import Exposures, INDICATOR_IMPF
@@ -153,12 +154,11 @@ https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/DHXBJX
             i_1 = 7  # get sum over all crops (columns 7 to 48)
             i_2 = 49
         self.gdf['value'] = data.iloc[:, i_1:i_2].sum(axis=1).values
-        self.gdf['latitude'] = lat.values
-        self.gdf['longitude'] = lon.values
+        self.gdf['geometry'] = [Point(x,y) for x,y in zip(lon.values, lat.values)]
         LOGGER.info('Lat. range: {:+.3f} to {:+.3f}.'.format(
-            np.min(self.gdf.latitude), np.max(self.gdf.latitude)))
+            np.min(self.gdf.geometry.y), np.max(self.gdf.geometry.y)))
         LOGGER.info('Lon. range: {:+.3f} to {:+.3f}.'.format(
-            np.min(self.gdf.longitude), np.max(self.gdf.longitude)))
+            np.min(self.gdf.geometry.x), np.max(self.gdf.geometry.x)))
 
         # set region_id (numeric ISO3):
         country_id = data.loc[:, 'iso3']
@@ -189,7 +189,7 @@ https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/DHXBJX
             self.value_unit = 'USD'
 
         LOGGER.info('Total {} {} {}: {:.1f} {}.'.format(
-            spam_v, spam_t, region, self.gdf.value.sum(), self.value_unit))
+            spam_v, spam_t, region, self.value.sum(), self.value_unit))
         self.check()
 
     def _set_impf(self, spam_t, haz_type):
@@ -218,7 +218,6 @@ https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/DHXBJX
             self.description += "\nrainfed portion of crop (= TA - TI)"
         else:
             self.gdf[INDICATOR_IMPF + haz_type] = 1
-        self.set_geometry_points()
 
     def _read_spam_file(self, **parameters):
         """Reads data from SPAM CSV file and cuts out the data for the