+++ climada_petals/hazard/tc_surge_geoclaw/__init__.py
@@ -13,32 +13,7 @@ PARTICULAR PURPOSE.  See the GNU General Public License for more details.
 
 You should have received a copy of the GNU General Public License along
 with CLIMADA. If not, see <https://www.gnu.org/licenses/>.
-
----
-
-Test files_handler module.
 """
 
-from pathlib import Path
-import unittest
-
-import pytest
-
-from climada_petals.hazard.tc_tracks_forecast import TCForecast
-
-
-class TestDataAvail(unittest.TestCase):
-    """Test availability of data used through APIs"""
-
-    def test_ecmwf_tc_bufr(self):
-        """Test availability ECMWF essentials TC forecast."""
-        fcast = TCForecast.fetch_bufr_ftp()
-        [f.close() for f in fcast]
-
-
-# Execute Tests
-if __name__ == '__main__':
-    from sys import argv
-
-    outputdir = argv[1] if len(argv) > 1 else str(Path.cwd().joinpath("tests_xml"))
-    pytest.main([f"--junitxml={outputdir}/tests.xml", __file__])
+from .sea_level_funs import *
+from .tc_surge_geoclaw import *