+++ climada_petals/test/test_cropyield_integr.py
@@ -23,6 +23,7 @@ import unittest
 import numpy as np
 from climada.engine import Impact
 from climada.entity import ImpactFuncSet
+from climada.entity.exposures import INDICATOR_IMPF
 from climada.util.constants import DEMO_DIR as INPUT_DIR
 from climada_petals.entity import ImpfRelativeCropyield
 from climada_petals.entity.exposures.crop_production import CropProduction, value_to_usd
@@ -52,6 +53,7 @@ class TestIntegr(unittest.TestCase):
                                                 crop='whe', irr='firr')
         exp = value_to_usd(exp, INPUT_DIR, yearrange=(2000, 2018))
         exp.assign_centroids(haz, threshold=20)
+        exp.data[INDICATOR_IMPF + haz_new.haz_type] = 1
 
         impf_cp = ImpactFuncSet()
         impf_def = ImpfRelativeCropyield.impf_relativeyield()
@@ -60,7 +62,7 @@ class TestIntegr(unittest.TestCase):
 
         impact = Impact()
         reg_sel = exp.copy()
-        reg_sel.gdf = reg_sel.gdf[reg_sel.gdf.region_id == 276]
+        reg_sel.data = reg_sel.gdf[reg_sel.gdf.region_id == 276]
         impact.calc(reg_sel, impf_cp, haz_new.select(['2002']), save_mat=True)
 
         exp_manual = reg_sel.gdf.value
@@ -71,10 +73,10 @@ class TestIntegr(unittest.TestCase):
         self.assertEqual(haz_new.size, 5)
         self.assertEqual(haz_new.centroids.size, 1092)
         self.assertAlmostEqual(haz_new.intensity.mean(), -2.0489097e-08, places=0)
-        self.assertAlmostEqual(exp.gdf.value.max(), 52278210.72839116, places=0)
-        self.assertEqual(exp.gdf.latitude.values.size, 1092)
-        self.assertAlmostEqual(exp.gdf.value[3], 0.0)
-        self.assertAlmostEqual(exp.gdf.value[1077], 398947.79657832277, places=0)
+        self.assertAlmostEqual(exp.value.max(), 52278210.72839116, places=0)
+        self.assertEqual(exp.latitude.size, 1092)
+        self.assertAlmostEqual(exp.value[3], 0.0)
+        self.assertAlmostEqual(exp.value[1077], 398947.79657832277, places=0)
         self.assertAlmostEqual(impact.imp_mat.data[3], -178745.59091285995, places=0)
         self.assertEqual(len(dif), 0)
 
@@ -93,6 +95,7 @@ class TestIntegr(unittest.TestCase):
                                                 bbox=bbox, yearrange=(2001, 2005),
                                                 scenario='flexible', unit='t/y', crop='whe', irr='firr')
         exp.assign_centroids(haz, threshold=20)
+        exp.data[INDICATOR_IMPF + haz.haz_type] = 1
 
         impf_cp = ImpactFuncSet()
         impf_def = ImpfRelativeCropyield.impf_relativeyield()
@@ -107,6 +110,7 @@ class TestIntegr(unittest.TestCase):
                                               scenario='flexible', unit='t/y', crop='whe', irr='firr')
         exp_nan.gdf.value[exp_nan.gdf.value==0] = np.nan
         exp_nan.assign_centroids(haz, threshold=20)
+        exp_nan.data[INDICATOR_IMPF + haz.haz_type] = 1
 
         impact_nan = Impact()
         impact_nan.calc(exp_nan, impf_cp, haz, save_mat=True)